requirements.txt
numpy
Cython
pycddlib
python-qsoptex